% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tennis.R
\name{get_score}
\alias{get_score}
\title{get_score}
\usage{
get_score(tennis_game)
}
\arguments{
\item{tennis_game}{a tennis_game}
}
\value{
a string
}
\description{
get_score
}
\examples{
tennis_game |>
won_point(playerName = "player1") |>
won_point(playerName = "player2") |>
get_score()
}
