% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tennis.R
\name{won_point}
\alias{won_point}
\title{playerName won a point}
\usage{
won_point(tennis_game, playerName)
}
\arguments{
\item{tennis_game}{a tennis_game}

\item{playerName}{player name as string}
}
\value{
a tennis_game
}
\description{
playerName won a point
}
\examples{

tennis_game |>
 won_point(playerName = "player1")
}
